scanner: consistently +2 to realloc size for EOB
We expect to have room for 2 EOB chars, but in this particular realloc
path this only works because yy_n_chars tends to be large enough for
1.5x its value to have room for our 2 EOB chars.

That said, clang-6's static analyzer does produce a warning about this
("Use of zero-allocated memory", meaning realloc with size=0) because it
goes by the assumption that yy_n_chars and number_to_move could be 0.

While this is somewhat bogus, it does make sense to just put the "+ 2"
here and be done with it.  AFAICS, this is the correct thing to do here
after all.

Signed-off-by: David Lamparter <equinox@diac24.net>